fda-analyse-service/requirements.txt
messytables==0.15.2
pandas==1.2.3
flask==1.1.2
flasgger==0.9.5
psycopg2-binary==2.8.6
py-eureka-client==0.9.1
docker==5.0.0
sqlalchemy==1.4.15
